Fix error object when a service throws
The error object needs to be in an "error" field. The code above does this correctly (Method not found) but this code doesn't. The result is that if an unhandled exception occurred the client would get a ServiceUnregistered event and then "ServiceDisappeared" responses. With this, it gets the expected error and does not unregister the service.