Changed type of property library in Class from @Object to @Library.
This property should never be an Instance and should probably never be
null. If that's the case we'll need a proper fix, but we'll need more
information to determine what's happening. If library is null, it should
now be handled and converted to null instead of Instance::null, although
this should never be the case and is likely a bug.